Crate p256_cortex_m4_sys

Structs

SignPrecomp

Sign precomputation state.

__fsid_t

max_align_t

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__LONG_DOUBLE_USES_FLOAT128

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_ISO_10646__

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE32_PTRDIFF_LONG

__WORDSIZE32_SIZE_ULONG

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

false_

has_d_cache

has_fpu

include_p256_decode_point

include_p256_decompress_point

include_p256_ecdh

include_p256_keygen

include_p256_raw_scalarmult_base

include_p256_raw_scalarmult_generic

include_p256_sign

include_p256_to_octet_string_compressed

include_p256_to_octet_string_hybrid

include_p256_to_octet_string_uncompressed

include_p256_verify

true_

use_fast_p256_basemult

use_mul_for_sqr

Functions

P256_check_range_n⚠

Checks that the argument, as little-endian integer, is a reduced non-zero element of the scalar field.

P256_check_range_p⚠

Checks that the argument, as little-endian integer, is a reduced element of the base field.

p256_convert_endianness⚠

Converts endianness by reversing the input value.

p256_ecdh_calc_shared_secret⚠

Generates the shared secret according to the ECDH standard.

p256_keygen⚠

Calculates the public key from a given private key for use by either ECDSA or ECDH.

p256_octet_string_to_point⚠

Decodes a point according to the three encodings above.

p256_point_to_octet_string_compressed⚠

Compressed encoding: “02 || Px” if Py is even and “03 || Px” if Py is odd.

p256_point_to_octet_string_hybrid⚠

Hybrid encoding: “06 || Px || Py” if Py is even and “07 || Px || Py” if Py is odd (a pretty useless encoding).

p256_point_to_octet_string_uncompressed⚠

Uncompressed encoding: “04 || Px || Py”.

p256_scalarmult_base⚠

Raw scalar multiplication by the base point of the elliptic curve.

p256_scalarmult_generic⚠

Raw scalar multiplication by any point on the elliptic curve.

p256_sign⚠

Creates an ECDSA signature.

p256_sign_step1⚠

Creates an ECDSA signature, using a two-step procedure.

p256_sign_step2⚠

Second step of creating an ECDSA signature, using a two-step procedure.

p256_verify⚠

Verifies an ECDSA signature.

Type Definitions

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time64_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

size_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t